fix trait definitions and update generics
This change adds
1. Stepper trait definition
2. Stepper generics for the Lcpp28byj to abstract pins
3. Better naming between Traits + Specific devices that implement those
   traits
